Revert "TAP5-2333: Optimize OperationTracker logging for normal (success, no logging) case"
On further reflection, I realized that this change behaves properly only for test cases
where the OperationException is allowed to propogate all the way back to a "top level" and
be reported there. In actual practice, there will be other layers intercepting the exception
and consuming it. Further, some of these layers (such as the ExceptionReport page logic)
requires the complete operation trace, not the partial trace available while unwinding.

This reverts commit 1cb63a12b4887b54210d91fe4a9863cf3dad93b5.