TAP5-2338: Work around some IE8 wierdness
The T5 floating console was always visible since there was an error trying to write to the native console, which simply does not exist in IE8
unless the developer tools are enabled. This fix drops attempts to write to the native console when it is missing. It is then possible to dismiss the
floating console normally.