TAP5-2661: GoogleClosureMinimizerOptionsProvider added
The Google Closure compiler was previously used with a fixed options. Due to JavaScript edge-cases this could lead to problems for some JS code.

So far the only solutions were to either disable compression completely, or set the Request parameter "tapestry.disable-javascript-minimization", for example with a RequestFilter.

To give the user more fine-grained control over the options, the "GoogleClosureMinimizerOptionsProvider" is added. In addition to provide custom compiler options, an empty Optional can be returned, and the StreamableResource will be transferred "as-is".