disable tapestry-javadoc on Java 8
The Java 8 build was still failing, and this is not a runtime requirement.
Due to changes in the Java tooling between 8 and 11 tapestry-javadoc is no longer compatible with 8.
So now all gradle related calls check for != 8 before using it.
There still might be issues if build with Java 9, but right know Travis only uses 8 and 11.