webresources: un-ignored TypeScriptCompilerSpec
The TypeScriptCompilerSpec was ignored because it used Tapestry to build itself, and there were some Spock issues.
To get it working, the TypeScriptCompiler is now build manually.

A test for TAP5-2691 (Rhino problems) was added. It succeeds on my machine, but maybe not on others. As long as the issue isn't actively worked on, the test is ignored.