TAP5-2708: Updated CssCompressor, fix errors, and create Spec with original tests
The CssCompressor in tapestry-webresources is now based on the latest YUI commit cf0497e.
To make it work some errors had to be fixed:
- Duplicate variable declaration of "oldCss"
- Faulty non-escaped regex
- zeros.css test failed

I'm not 100% sure the CssCompressor behaves 100% as before, but it's a necessary trade-off so it can be improved further, as the original YUI project is dead.