TAP5-2763: improve @RequestParameter blank handling
@RequestParamter now also has "boolean treatBlankAsNull() default false" to better handl blank values.

A value might be blank, even if it's supposed to be a null coming from the client, as any parameter in a request set to null will end up as blank on the server side.
This might lead to coercion problems if no sensible coercer from blank to the target type exists.
To keep the handling as close to the original, an additional annotation parameter made more sense than changing the overall handling.